rename : HomePage를 MainPage로 변경 및 ExPage 삭제
