--- client/src/pages/ExPage/ExPage.styles.tsx
@@ -1,7 +1,0 @@
-import styled from 'styled-components'
-
-export const Container = styled.div`
-  width: 100vw;
-  height: 100vh;
-  background-color: #b1b1b1;
-`
