--- client/src/pages/ExPage/ExPage.tsx
@@ -1,11 +1,0 @@
-import * as styles from './ExPage.styles'
-
-const ExPage = () => {
-  return (
-    <styles.Container>
-      <h1>exPage</h1>
-    </styles.Container>
-  )
-}
-
-export default ExPage
