+++ client/src/pages/MainPage/MainPage.styles.tsx
